Generalize the audio classification event
The audio classification event tns1:AudioAnalytics/Audio/DetectedSound
includes an AudioClassDescriptor which includes 0 or more
AudioClassCandidate. AudioClassCandidate has a Type and Likelihood but
the Type is an enum which could be problematic to extend because some
XML parsers would break. So this limits us to only detect the types
that are defined in the enum: gun_shot, scream, glass_breaking and
tire_screetch. Add a new rule and event to fix this issue which has
a StringList instead with detected classes. Predefined audio classes
are listed in tt:AudioClassification.

Change-Id: I40558209d78ab11b54b68e572c7adb0a00646bcc